Update PythonVirtualEnvironment
This script will now create a virtual environment using the specified
Python Interpreter. It wil then install the dependencies specified
in the requirements.txt, which can be a single file or multiple.

It will add the same environmental variable used by python activate
to an activation script named after the Conan package.

Usage in a conanfile.py
```python
class ProjectConan(conanfile)
...
    name = "superawesomeproject"
    python_requires = "PythonVirtualEnvironment/0.2.0@ultimaker/testing"

    @property
    def _site_packages(self):
        if self.settings.os == "Windows":
            return os.path.join("Lib", "site-packages")
        return os.path.join("lib", "python3.10", "site-packages")

    def layout(self):
        self.folders.source = "."
        self.folders.build = "venv"
        self.folders.generators = os.path.join("venv", "bin")

    def generate(self):
        venv = self.python_requires["PythonVirtualEnvironment"].module.PythonVirtualEnvironment(self)
        venv.configure(python_interpreter = sys.executable, requirements_txt = ("requirements.txt", "requirements-dev.txt"))
        venv.generate()
```

Calling the Conan install command on a project with the above-shown
code:

```bash
conan install .
```

Will create a virtual environment using the system python interpreter,
it will install all requirements in the two requirements files (relative
to the source folder) and it will add all run information from other
packages to the environment. Such as Arcus and all other dependencies.

The environment can be activated with:
```
# Linux / MacOS
. venv/bin/superawesomeproject.sh

# Windows cmd
./venv/Scripts/superawesomeproject.bat

# Windows powershell
./venv/Scripts/superawesomeproject.ps1
```

once the environment is activated you can run python:

```
$ python
>>> import pyArcus
>>> import PyQt6
```

Deactivation is done with:

```
# Linux / MacOS
. venv/bin/deactivate_superawesomeproject.sh

# Windows cmd
./venv/Scripts/deactivate_superawesomeproject.bat

# Windows powershell
./venv/Scripts/deactivate_superawesomeproject.ps1
```

Contributes to CURA-9177