Fixed compiling on Windows
Disbabled harfbuzz and glib for now since these require the use of a bash such as msys2. If we still need
these libraries. We should fix these recipes such that they don't use automake and autoconf on Windows and
add the configure option: `-o msys2:packages=base-devel,binutils,gcc,autoconf,automake`
These recipes are older version and don't handle the the run/build environment and the win_bash config options
well. Preinstalling these packages is a quick and dirty solution but a viable one due to the time constraints.

Contribute to CURA-8831