Fixed linking against Py_LIMITED_API
Might be a bug in PyQt-builder but the option link-full-dll isn't available, even though it is set in the
module pyqtbuild\project.py. A lot of sip definition files set the `use_limited_api` to `True` for the module
Since we compile PyQt6 for a single application against our own Python version, we can set the
`use_limited_api` to `False` this should stop the linker from linking against the limited Python ABI
python3.lib. Because now the `Py_LIMITED_API` C preprocessor symbol is no longer used.
Keep in mind that this is a quick and dirty hack, which allows us to link against the full
python<major><minor> ABI.

Contributes to CURA-8831